Alterado o tipo do ID de int para Integer para permitir o ToString

+++ src/main/java/br/com/fatesg/fabrica/projetofabrica/Banco.java
@@ -21,7 +21,7 @@ public class Banco implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	@Getter @Id @GeneratedValue(strategy=GenerationType.AUTO) 
-	private Long cod;     
+	private Integer cod;     
     
 	@NotNull 
 	private String banco;
